Added the fix for broken Data Template window
