+++ src/renderer/components/Data/DatasetPreviewWithTemplate.tsx
@@ -1,16 +1,18 @@
-import { useEffect, useState, useRef, useCallback } from 'react';
+import { useEffect, useState } from 'react';
 import {
+  Button,
   Table,
-  Input,
   CircularProgress,
   Box,
+  IconButton,
+  iconButtonClasses,
   Alert,
+  Chip,
   LinearProgress,
-  Button,
-  Typography,
 } from '@mui/joy';
 
 import * as chatAPI from '../../lib/transformerlab-api-sdk';
+import { ChevronLeftIcon, ChevronRightIcon, Sheet } from 'lucide-react';
 import useSWR from 'swr';
 const fetcher = (url) =>
   fetch(url)
@@ -18,182 +20,209 @@ const fetcher = (url) =>
     .then((data) => data);
 
 const DatasetTableWithTemplate = ({ datasetId, template }) => {
-  const [rows, setRows] = useState([]);
-  const [columns, setColumns] = useState([]);
-  const [offset, setOffset] = useState(0);
-  const [hasMore, setHasMore] = useState(true);
-  const [loading, setLoading] = useState(false);
-  const [searchText, setSearchText] = useState('');
-  const [saving, setSaving] = useState(false);
-  const [modifiedRows, setModifiedRows] = useState(new Map());
-  const limit = 50;
-  const containerRef = useRef(null);
-
-  const loadMore = useCallback(async () => {
-    if (loading || !hasMore) return;
-    setLoading(true);
-    try {
-      const url = chatAPI.Endpoints.Dataset.PreviewWithTemplate(
-        datasetId,
-        encodeURIComponent(template),
-        offset,
-        limit,
-      );
-      const result = await fetcher(url);
-      if (result.status === 'success') {
-        const newRows = result.data.rows || [];
-        setRows((prev) => [...prev, ...newRows]);
-        setColumns(result.data.columns || []);
-        setOffset((prev) => prev + limit);
-        if (newRows.length < limit) setHasMore(false);
-      } else {
-        setHasMore(false);
-      }
-    } catch (e) {
-      setHasMore(false);
-    } finally {
-      setLoading(false);
-    }
-  }, [datasetId, template, offset, loading, hasMore]);
-
-  useEffect(() => {
-    setRows([]);
-    setColumns([]);
-    setOffset(0);
-    setHasMore(true);
-    setLoading(false);
-    setModifiedRows(new Map());
-  }, [datasetId, template]);
-
-  useEffect(() => {
-    loadMore();
-  }, [loadMore]);
-
-  const onScroll = (e) => {
-    const { scrollTop, scrollHeight, clientHeight } = e.target;
-    if (scrollHeight - scrollTop <= clientHeight + 50) {
-      loadMore();
-    }
+  const [pageNumber, setPageNumber] = useState(1);
+  const [numOfPages, setNumOfPages] = useState(1);
+  const [datasetLen, setDatasetLen] = useState(null);
+  let pageSize = 4; //Set the number of rows per page
+  const offset = (pageNumber - 1) * pageSize; //Calculate current row number to start from
+  //Set the pagination for the dataset
+  const setPagination = (totalRows, rowsPerPage) => {
+    const totalPages = Math.ceil(totalRows / rowsPerPage);
+    setNumOfPages(totalPages);
   };
-
-  const imageKey = columns.find((key) => key.toLowerCase().includes('image'));
-  const textKey = columns.find(
-    (key) =>
-      key.toLowerCase().includes('text') ||
-      key.toLowerCase().includes('caption'),
+  const {
+    data: result,
+    error,
+    isLoading,
+    mutate,
+  } = useSWR(
+    chatAPI.Endpoints.Dataset.PreviewWithTemplate(
+      datasetId,
+      encodeURIComponent(template),
+      offset,
+      pageSize,
+    ),
+    fetcher,
   );
 
-  const updateCaption = (index, newText) => {
-    setRows((prev) => {
-      const updated = [...prev];
-      updated[index] = { ...updated[index], [textKey]: newText };
-      const uniqueKey = updated[index]['__index__'];
-      setModifiedRows((prevMap) =>
-        new Map(prevMap).set(uniqueKey, updated[index]),
-      );
-      return updated;
-    });
-  };
+  useEffect(() => {
+    setDatasetLen(null);
+    setPageNumber(1);
+    setNumOfPages(1);
+  }, [datasetId]);
 
-  const filteredRows = rows.filter((row) => {
-    const text = row[textKey] || '';
-    return text.toLowerCase().includes(searchText.toLowerCase());
-  });
-
-  const saveEdits = async () => {
-    setSaving(true);
-    try {
-      // Create FormData containing the modified rows
-      const formData = new FormData();
-      const blob = new Blob(
-        [JSON.stringify(Array.from(modifiedRows.values()))],
-        { type: 'application/json' },
-      );
-      formData.append('file', blob, 'metadata_updates.json'); // singular, not "files"
-      formData.append('dataset_id', datasetId); // required by backend as Form(...)
-
-      // Send POST request to the same style endpoint
-      const response = await fetch(
-        chatAPI.Endpoints.Dataset.SaveMetadata(datasetId),
-        {
-          method: 'POST',
-          body: formData,
-        },
-      );
-
-      if (!response.ok) throw new Error('Failed to save');
-      alert('Captions saved successfully!');
-    } catch (err) {
-      alert('Error saving captions');
-    } finally {
-      setSaving(false);
+  useEffect(() => {
+    if (result && result.data && datasetLen === null) {
+      setDatasetLen(result?.data['len']);
+      setPagination(result?.data['len'], pageSize);
     }
-  };
+  }, [result, pageSize, datasetLen]);
 
+  if (!result?.data?.rows) {
+    if (isLoading) {
+      return <LinearProgress />;
+    }
+    return '';
+  }
   return (
-    <Box
-      ref={containerRef}
-      onScroll={onScroll}
-      sx={{
-        overflow: 'auto',
-        height: '100%',
-        display: 'flex',
-        flexDirection: 'column',
-      }}
-    >
-      {loading && rows.length === 0 && <LinearProgress />}
-      <Box p={1} display="flex" gap={2} alignItems="center">
-        <Input
-          placeholder="Search captions..."
-          fullWidth
-          value={searchText}
-          onChange={(e) => setSearchText(e.target.value)}
-        />
-        <Button onClick={saveEdits} loading={saving} variant="soft">
-          Save Changes
-        </Button>
+    <>
+      {/* <pre>{JSON.stringify(data, null, 2)}</pre> */}
+      <Box sx={{ overflow: 'auto', height: '100%' }}>
+        {isLoading && <LinearProgress />}
+        {result?.status == 'error' && (
+          <Alert color="danger">{result?.message}</Alert>
+        )}
       </Box>
-      <Box sx={{ overflow: 'auto', flex: 1 }}>
-        <Table sx={{ minWidth: '100%' }}>
+      {result?.data && (
+        <Table sx={{ tableLayout: 'auto', overflow: 'scroll' }}>
           <thead>
             <tr>
-              <th>Image</th>
-              <th>Description</th>
+              <th>Rendered (this is what the model sees)</th>
+              <th>Fields</th>
+              {/* {Object.keys(data.data[0]).map((key) => {
+                if (key.startsWith('__') && key.endsWith('__')) {
+                  return null;
+                } else {
+                  return <th key={key}>{key}</th>;
+                }
+              })} */}
             </tr>
           </thead>
           <tbody>
-            {filteredRows.map((row, idx) => (
-              <tr key={row?.['__index__'] || idx}>
+            {result?.data?.rows?.map((row) => (
+              <tr key={row?.['__index__']}>
                 <td>
-                  {imageKey &&
-                    row[imageKey] &&
-                    typeof row[imageKey] === 'string' && (
-                      <img
-                        src={row[imageKey]}
-                        alt={`example-${idx}`}
-                        style={{ maxHeight: '100px' }}
-                      />
-                    )}
+                  <pre style={{ whiteSpace: 'pre-wrap' }}>
+                    <Box
+                      style={{
+                        backgroundColor: 'var(--joy-palette-success-100)',
+                        padding: '5px',
+                      }}
+                    >
+                      {row?.['__formatted__']}
+                    </Box>
+                  </pre>
                 </td>
+                {/* {Object.entries(row).map(([key, value]) => {
+                  // if key starts and ends with __ then skip:
+                  if (key.startsWith('__') && key.endsWith('__')) {
+                    return null;
+                  } else {
+                    return <td>{value}</td>;
+                  }
+                })} */}
                 <td>
-                  <Input
-                    value={row[textKey] || ''}
-                    onChange={(e) => updateCaption(idx, e.target.value)}
-                    onKeyDown={(e) => {
-                      if (e.key === 'Enter') e.currentTarget.blur();
-                    }}
-                    variant="soft"
-                    size="sm"
-                    fullWidth
-                  />
+                  <pre style={{ whiteSpace: 'pre-wrap' }}>
+                    {/* {JSON.stringify(row, null, 2)} */}
+                    {Object.entries(row).map(([key, value]) => {
+                      if (key.startsWith('__') && key.endsWith('__')) {
+                        return null;
+                      } else {
+                        return (
+                          <Box>
+                            {typeof value === 'string' ? (
+                              <>
+                                <Chip>{key}</Chip>
+                                {value?.length > 200 ? (
+                                  <>
+                                    {value.substring(0, 200)}
+                                    ...
+                                  </>
+                                ) : (
+                                  value
+                                )}
+                              </>
+                            ) : (
+                              <pre>{JSON.stringify(value, null, 2)}</pre>
+                            )}
+                          </Box>
+                        );
+                      }
+                    })}
+                  </pre>
                 </td>
               </tr>
             ))}
           </tbody>
         </Table>
+      )}
+      <Box
+        className="Pagination"
+        sx={{
+          pt: 2,
+          gap: 1,
+          [`& .${iconButtonClasses.root}`]: { borderRadius: '50%' },
+          display: 'inline-flex',
+          alignItems: 'center',
+        }}
+      >
+        {pageNumber > 1 ? (
+          <Button
+            size="sm"
+            variant="outlined"
+            color="neutral"
+            onClick={() => setPageNumber(pageNumber - 1)}
+          >
+            <ChevronLeftIcon /> Previous
+          </Button>
+        ) : (
+          <div style={{ width: '78px', height: '30px' }} />
+        )}
+        <Box sx={{ flex: 1, alignItems: 'center' }} />
+        <IconButton
+          key={1}
+          size="sm"
+          variant={Number(1) === pageNumber ? 'outlined' : 'plain'}
+          color="neutral"
+          onClick={() => setPageNumber(Number(1))}
+        >
+          {1}
+        </IconButton>
+        {pageNumber > 4 ? '…' : <div />}
+        {Array.from(
+          { length: Math.min(5, numOfPages) },
+          (_, i) => pageNumber + i - 2,
+        )
+          .filter((page) => page >= 2 && page < numOfPages)
+
+          .map((page) => (
+            <IconButton
+              key={page}
+              size="sm"
+              variant={page === pageNumber ? 'outlined' : 'plain'}
+              color="neutral"
+              onClick={() => setPageNumber(Number(page))}
+            >
+              {page}
+            </IconButton>
+          ))}
+        {pageNumber < numOfPages - 4 ? '…' : <div />}
+        {numOfPages != 1 && (
+          <IconButton
+            key={numOfPages}
+            size="sm"
+            variant={Number(numOfPages) === pageNumber ? 'outlined' : 'plain'}
+            color="neutral"
+            onClick={() => setPageNumber(Number(numOfPages))}
+          >
+            {numOfPages}
+          </IconButton>
+        )}
+        <Box sx={{ flex: 1 }} />
+        {pageNumber < numOfPages ? (
+          <Button
+            size="sm"
+            variant="outlined"
+            color="neutral"
+            onClick={() => setPageNumber(pageNumber + 1)}
+          >
+            Next <ChevronRightIcon />
+          </Button>
+        ) : (
+          <div style={{ width: '78px', height: '30px' }} />
+        )}
       </Box>
-      {loading && rows.length > 0 && <LinearProgress />}
-    </Box>
+    </>
   );
 };
